+++ src/mix_set_readable.d.ts
@@ -5,3 +5,10 @@ export interface mix_set_readable_I<Val extends unknown = unknown> extends Reada
     set: Writable_set_T<Val>;
     update: Writable_update_T<Val>;
 }
+export declare class mix_set_readable_C<Val extends unknown = unknown> implements mix_set_readable_I<Val> {
+    protected store: mix_set_readable_I<Val>;
+    constructor(store: mix_set_readable_I<Val>);
+    readonly subscribe: (this: void, run: import("svelte/store").Subscriber<Val>, invalidate?: ((value?: Val | undefined) => void) | undefined) => import("svelte/store").Unsubscriber;
+    readonly set: Writable_set_T<Val>;
+    readonly update: Writable_update_T<Val>;
+}