+++ src/mix_set_readable.ts
@@ -16,3 +16,9 @@ export interface mix_set_readable_I<Val extends unknown = unknown> extends Reada
 	set:Writable_set_T<Val>
 	update:Writable_update_T<Val>
 }
+export class mix_set_readable_C<Val extends unknown = unknown> implements mix_set_readable_I<Val> {
+	constructor(protected store:mix_set_readable_I<Val>) {}
+	readonly subscribe = this.store.subscribe
+	readonly set = this.store.set
+	readonly update = this.store.update
+}